+++ configs/pymafx/pymafx.py
@@ -1,5 +1,5 @@
 _base_ = ['../_base_/default_runtime.py']
-__maf_on__ = True
+maf_on = True
 __bhf_mode__ = 'full_body'  # full_body or body_hand
 __grid_align__ = dict(
     use_att=True,
@@ -57,7 +57,7 @@ __hf_model_cfg__ = dict(
             num_deconv_filters=(256, 256, 256),
             num_deconv_kernels=(4, 4, 4),
             num_layers=50),
-        global_mode=not __maf_on__), )
+        global_mode=not maf_on), )
 
 __mesh_model__ = dict(
     name='smplx',
@@ -69,11 +69,11 @@ model = dict(
     backbone=dict(
         type='PoseHighResolutionNetPyMAFX',
         extra=__hrnet_extra__,
-        global_mode=not __maf_on__,
+        global_mode=not maf_on,
     ),
     head=dict(
         type='PyMAFXHead',
-        maf_on=__maf_on__,
+        maf_on=maf_on,
         n_iter=3,
         bhf_mode=__bhf_mode__,
         grid_feat=False,
@@ -93,10 +93,11 @@ model = dict(
     smplx_model_dir='data/body_models/smplx',
     mesh_model=__mesh_model__,
     bhf_mode=__bhf_mode__,
-    maf_on=__maf_on__,
+    maf_on=maf_on,
     body_sfeat_dim=[192, 96, 48],
     hf_sfeat_dim=(256, 256, 256),
     grid_feat=False,
+    aux_supv_on=maf_on,
     grid_align=__grid_align__,
     mlp_dim=[256, 128, 64, 5],
     hf_mlp_dim=[256, 128, 64, 5],
