--- mmhuman3d/models/bert/__init__.py
@@ -1,1 +1,0 @@
-__version__ = '1.0.0'
