+++ mmhuman3d/models/heads/bert/modeling_utils.py
@@ -27,19 +27,6 @@ from torch import nn
 
 logger = logging.getLogger(__name__)
 
-try:
-    from torch.nn import Identity
-except ImportError:
-    # Older PyTorch compatibility
-    class Identity(nn.Module):
-        """A placeholder identity operator that is argument-insensitive."""
-
-        def __init__(self, *args, **kwargs):
-            super(Identity, self).__init__()
-
-        def forward(self, input):
-            return input
-
 
 class PretrainedConfig(object):
     """Base class for all configuration classes.
@@ -475,31 +462,3 @@ class PreTrainedModel(nn.Module):
             return model, loading_info
 
         return model
-
-
-def prune_linear_layer(layer, index, dim=0):
-    """Prune a linear layer (a model parameters) to keep only entries in index.
-
-    Return the pruned layer as a new layer with requires_grad=True. Used to
-    remove heads.
-    """
-    index = index.to(layer.weight.device)
-    W = layer.weight.index_select(dim, index).clone().detach()
-    if layer.bias is not None:
-        if dim == 1:
-            b = layer.bias.clone().detach()
-        else:
-            b = layer.bias[index].clone().detach()
-    new_size = list(layer.weight.size())
-    new_size[dim] = len(index)
-    new_layer = nn.Linear(
-        new_size[1], new_size[0], bias=layer.bias
-        is not None).to(layer.weight.device)
-    new_layer.weight.requires_grad = False
-    new_layer.weight.copy_(W.contiguous())
-    new_layer.weight.requires_grad = True
-    if layer.bias is not None:
-        new_layer.bias.requires_grad = False
-        new_layer.bias.copy_(b.contiguous())
-        new_layer.bias.requires_grad = True
-    return new_layer
